Use same exec implementation for SwiftRunCommand and DebugTestRunner
