+++ Sources/Commands/SwiftRunCommand.swift
@@ -19,7 +19,6 @@ import PackageModel
 import SPMBuildCore
 
 import enum TSCBasic.ProcessEnv
-import func TSCBasic.exec
 
 import enum TSCUtility.Diagnostics
 
@@ -188,7 +187,7 @@ public struct SwiftRunCommand: AsyncSwiftCommand {
                 } else {
                     let pathRelativeToWorkingDirectory = productAbsolutePath.relative(to: swiftCommandState.originalWorkingDirectory)
                     let lldbPath = try swiftCommandState.getTargetToolchain().getLLDB()
-                    try exec(path: lldbPath.pathString, args: ["--", pathRelativeToWorkingDirectory.pathString] + options.arguments)
+                    try safeExec(path: lldbPath.pathString, args: ["--", pathRelativeToWorkingDirectory.pathString] + options.arguments)
                 }
             } catch let error as RunError {
                 swiftCommandState.observabilityScope.emit(error)
@@ -303,7 +302,7 @@ public struct SwiftRunCommand: AsyncSwiftCommand {
         }
 
         let pathRelativeToWorkingDirectory = executablePath.relative(to: originalWorkingDirectory)
-        try execute(path: executablePath.pathString, args: [pathRelativeToWorkingDirectory.pathString] + arguments)
+        try safeExec(path: executablePath.pathString, args: [pathRelativeToWorkingDirectory.pathString] + arguments)
     }
 
     /// Determines if a path points to a valid swift file.
@@ -324,41 +323,6 @@ public struct SwiftRunCommand: AsyncSwiftCommand {
             absolutePath = try AbsolutePath(cwd, validating: path)
         }
         return fileSystem.isFile(absolutePath)
-    }
-
-    /// A safe wrapper of TSCBasic.exec.
-    private func execute(path: String, args: [String]) throws -> Never {
-        #if !os(Windows)
-        // Dispatch will disable almost all asynchronous signals on its worker threads, and this is called from `async`
-        // context. To correctly `exec` a freshly built binary, we will need to:
-        // 1. reset the signal masks
-        for i in 1..<NSIG {
-            signal(i, SIG_DFL)
-        }
-        var sig_set_all = sigset_t()
-        sigfillset(&sig_set_all)
-        sigprocmask(SIG_UNBLOCK, &sig_set_all, nil)
-
-        #if os(FreeBSD) || os(OpenBSD)
-        #if os(FreeBSD)
-        pthread_suspend_all_np()
-        #endif
-        closefrom(3)
-        #else
-        #if os(Android)
-        let number_fds = Int32(sysconf(_SC_OPEN_MAX))
-        #else
-        let number_fds = getdtablesize()
-        #endif /* os(Android) */
-
-        // 2. set to close all file descriptors on exec
-        for i in 3..<number_fds {
-            _ = fcntl(i, F_SETFD, FD_CLOEXEC)
-        }
-        #endif /* os(FreeBSD) || os(OpenBSD) */
-        #endif
-
-        try TSCBasic.exec(path: path, args: args)
     }
 
     public init() {}
